Use stores in CloudFormation
