+++ localstack/services/cloudformation/models/__init__.py
@@ -25,4 +25,5 @@ __all__ = [
     "sqs",
     "ssm",
     "stepfunctions",
+    "stores",
 ]
